Fix Android compile error: setState() takes int, not long
PlaybackStateCompat.Builder.setState(int state, long position, float
speed) — the playback state constant was declared as long.

+++ selexiones-app/android/app/src/main/java/com/selexiones/app/RadioPlaybackService.java
@@ -160,7 +160,7 @@ public class RadioPlaybackService extends Service {
     }
 
     private void onPlaybackChanged(boolean playing) {
-        long state = playing ? PlaybackStateCompat.STATE_PLAYING : PlaybackStateCompat.STATE_PAUSED;
+        int state = playing ? PlaybackStateCompat.STATE_PLAYING : PlaybackStateCompat.STATE_PAUSED;
         mediaSession.setPlaybackState(
             new PlaybackStateCompat.Builder()
                 .setActions(PlaybackStateCompat.ACTION_PLAY | PlaybackStateCompat.ACTION_PAUSE | PlaybackStateCompat.ACTION_STOP)
